+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+### Features
+
+- Allow starting the SDK with an initial scope (#2982)
 - Swift Error Names (#2960)
 
 ```Swift